requirements.txt
paho-mqtt
pyyaml
pyserial
os
time
json
